Audit of integrations links and sidebar labels
Going through each link from the integrations-basics page, to make sure
the links are as correct as they can be, the labels in the sidebars
include the word "Integration", and fixing a few mini typos. There
are still a couple blank docs, which I'll fix as soon as I can.