Clarity on api
This change is editing the desc for this endpoint: /rest/v1.2/users/{username}/concurrency
The original desc stated you were getting VM and RDC but you do not get any RDC values and it is always 0. 
The api for RDC which I've added is: https://api.eu-central-1.saucelabs.com/v1/rdc/concurrency which then requires basic auth to be passed as your sauce creds. 
I'm not sure why this isn't in our docs but you can reject it if necessary as there is another section which customers aren't finding and they state it is currently misleading