temporarily delete docs on Okta SSO SAML app
The new SSO app has not been yet published in Okta catalog because
the Okta team keeps asking for fixing documentation. Until the new app
is published in the Okta catalog, I want to hide this documentation
because it may be confusing for customers as they may accidently set up
the new SSO integration using the legacy app from the Okta catalog.