fix docs for the new Okta SAML app according to Okta requirements
I based the tutorial on adding the Sauce Labs SAML app from the Okta
catalog on the following okta docs:
* https://help.okta.com/en-us/Content/Topics/Apps/apps-add-applications.htm
* https://help.okta.com/en-us/Content/Topics/Apps/Apps_Overview_of_Managing_Apps_and_SSO.htm

I removed screenshots as they were taken from the old app and we do not
have access to the new SAML app since it has not been yet published.
Once the app is published, I will add the screenshots.